refactor to SettingsService in package settings
